Add base attributes to BaseUseCase and refactor apimethod
