+++ healthchain/base.py
@@ -1,8 +1,11 @@
 from abc import ABC, abstractmethod
-from typing import Dict
+from typing import Dict, List, Optional
+
+from healthchain.service.service import Service
+from healthchain.service.endpoints import Endpoint
 
 from .workflows import UseCaseType, Workflow
-from .service.endpoints import Endpoint
+from .apimethod import APIMethod
 
 
 class BaseClient(ABC):
@@ -38,6 +41,22 @@ class BaseUseCase(ABC):
     - the data it accepts (FHIR or CDA)
     - the format of the request it constructs (CDS Hook or NoteReader workflows)
     """
+
+    def __init__(
+        self,
+        service_api: Optional[APIMethod] = None,
+        service_config: Optional[Dict] = None,
+        service: Optional[Service] = None,
+        client: Optional[BaseClient] = None,
+    ) -> None:
+        self._service_api: APIMethod = service_api
+        self._service: Service = service
+        self._client: BaseClient = client
+
+        self.service_config: service_config = service_config
+        self.responses: List[Dict[str, str]] = []
+        self.sandbox_id = None
+        self.url = None
 
     @property
     @abstractmethod
